florist-paradise/src/main/java/za/ac/cput/service/ContactService.java
package za.ac.cput.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import za.ac.cput.domain.Contact;
import za.ac.cput.repository.ContactRepository;

import java.util.List;

@Service
public class ContactService implements IContactService {

    private  ContactRepository contactRepository;
    @Autowired
    public ContactService(ContactRepository contactRepository) {
        this.contactRepository = contactRepository;
    }

    @Override
    public Contact save(Contact contact) {
        return contactRepository.save(contact);
    }

    @Override
    public Contact read(String s) {
        return contactRepository.findById(s).orElse(null);
    }

    @Override
    public Contact update(Contact contact) {
        return contactRepository.save(contact);
    }

    @Override
    public boolean deleteById(String s) {
        contactRepository.deleteById(s);
        return !contactRepository.existsById(s);
    }
    @Override
    public List<Contact> getall() {
        return contactRepository.findAll();
    }
}
